Extract home page wish card out for extensibility
